# Conflicts:
#	public/css/app.css
#	public/css/appdark.css
#	public/css/landing.css
#	public/js/compose.chunk.e1f297b242137d23.js.LICENSE.txt
#	public/js/compose.js.LICENSE.txt
#	public/js/daci.chunk.6ef63b08c34d044e.js
#	public/js/discover~findfriends.chunk.86d446fc3f9ceff9.js
#	public/js/discover~memories.chunk.3f2bb3780996e975.js
#	public/js/discover~myhashtags.chunk.429fff1b0030aa98.js
#	public/js/discover~serverfeed.chunk.ae0480fc3d475f13.js
#	public/js/discover~settings.chunk.8ebbefcb731bd3f4.js
#	public/js/hashtag.js
#	public/js/home.chunk.32b6ac057e5d00f5.js
#	public/js/home.chunk.32b6ac057e5d00f5.js.LICENSE.txt
#	public/js/home.chunk.8bbc3c5c38dde66d.js.LICENSE.txt
#	public/js/landing.js
#	public/js/manifest.js
#	public/js/post.chunk.275761c7dec9af41.js
#	public/js/post.chunk.275761c7dec9af41.js.LICENSE.txt
#	public/js/post.chunk.803d8c9f68415936.js.LICENSE.txt
#	public/js/profile.chunk.f18d6551c434b139.js
#	public/js/profile.js
#	public/js/spa.js
#	public/js/status.js
#	public/js/stories.js
#	public/js/story-compose.js
#	public/js/timeline.js
#	public/js/vendor.js
#	public/mix-manifest.json

+++ app/Http/Controllers/BookmarkController.php
@@ -3,12 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Bookmark;
-use App\Status;
-use Auth;
-use Illuminate\Http\Request;
+use App\Services\AccountService;
 use App\Services\BookmarkService;
 use App\Services\FollowerService;
 use App\Services\UserRoleService;
+use App\Status;
+use Illuminate\Http\Request;
 
 class BookmarkController extends Controller
 {
@@ -25,15 +25,16 @@ class BookmarkController extends Controller
 
         $user = $request->user();
         $status = Status::findOrFail($request->input('item'));
-
-        abort_if($user->has_roles && !UserRoleService::can('can-bookmark', $user->id), 403, 'Invalid permissions for this action');
+        $account = AccountService::get($status->profile_id);
+        abort_if(isset($account['moved'], $account['moved']['id']), 422, 'Cannot bookmark or unbookmark a post from an account that has migrated');
+        abort_if($user->has_roles && ! UserRoleService::can('can-bookmark', $user->id), 403, 'Invalid permissions for this action');
         abort_if($status->in_reply_to_id || $status->reblog_of_id, 404);
-        abort_if(!in_array($status->scope, ['public', 'unlisted', 'private']), 404);
-        abort_if(!in_array($status->type, ['photo','photo:album', 'video', 'video:album', 'photo:video:album']), 404);
+        abort_if(! in_array($status->scope, ['public', 'unlisted', 'private']), 404);
+        abort_if(! in_array($status->type, ['photo', 'photo:album', 'video', 'video:album', 'photo:video:album']), 404);
 
-        if($status->scope == 'private') {
-            if($user->profile_id !== $status->profile_id && !FollowerService::follows($user->profile_id, $status->profile_id)) {
-                if($exists = Bookmark::whereStatusId($status->id)->whereProfileId($user->profile_id)->first()) {
+        if ($status->scope == 'private') {
+            if ($user->profile_id !== $status->profile_id && ! FollowerService::follows($user->profile_id, $status->profile_id)) {
+                if ($exists = Bookmark::whereStatusId($status->id)->whereProfileId($user->profile_id)->first()) {
                     BookmarkService::del($user->profile_id, $status->id);
                     $exists->delete();
 
@@ -51,7 +52,7 @@ class BookmarkController extends Controller
             ['status_id' => $status->id], ['profile_id' => $user->profile_id]
         );
 
-        if (!$bookmark->wasRecentlyCreated) {
+        if (! $bookmark->wasRecentlyCreated) {
             BookmarkService::del($user->profile_id, $status->id);
             $bookmark->delete();
         } else {
